requirements.txt
asgiref==3.5.2
attrs==21.4.0
autobahn==22.7.1
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
Django==3.2.16
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
hyperlink==21.0.0
idna==3.4
importlib-resources==5.8.0
incremental==22.10.0
inflection==0.5.1
jsonschema==4.6.0
mysqlclient==2.1.1
Pillow==9.3.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.1.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
uritemplate==4.1.1
zipp==3.8.0
zope.interface==5.5.2
